actually support aint contraction #644
