+++ src/02-tagger/03-contractions/01-negative.js
@@ -12,10 +12,29 @@ const irregulars = {
   // "ain't" is ambiguous for is/was
 }
 
-const checkNegative = function(term) {
+// either 'is not' or 'are not'
+const doAint = function(term, phrase) {
+  let terms = phrase.cache.terms || phrase.terms()
+  let index = terms.indexOf(term)
+  let before = terms.slice(0, index)
+  //look for the preceding noun
+  let noun = before.find(t => {
+    return t.tags.Noun
+  })
+  if (noun && noun.tags.Plural) {
+    return ['are', 'not']
+  }
+  return ['is', 'not']
+}
+
+const checkNegative = function(term, phrase) {
   //check named-ones
   if (irregulars.hasOwnProperty(term.clean) === true) {
     return irregulars[term.clean]
+  }
+  //this word needs it's own logic:
+  if (term.clean === `ain't` || term.clean === 'aint') {
+    return doAint(term, phrase)
   }
   //try it normally
   if (hasNegative.test(term.clean) === true) {
